+++ apps/web/src/app/layout.tsx
@@ -1,32 +1,9 @@
-import './global.css';
-import {
-  ServerCreateProvider,
-  ServerJoinedTeamProvider,
-  ServerStateProvider,
-  ServerModalProvider,
-} from '@meezy/ui/client';
-
-export const metadata = {
-  title: 'Welcome to web',
-  description: 'Generated by create-nx-workspace',
-};
+import { RootLayoutWrapper } from '@meezy/ui/client';
 
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return (
-    <html lang="en">
-      <body>
-        <ServerCreateProvider>
-          <ServerJoinedTeamProvider>
-            <ServerModalProvider>
-              <ServerStateProvider>{children}</ServerStateProvider>
-            </ServerModalProvider>
-          </ServerJoinedTeamProvider>
-        </ServerCreateProvider>
-      </body>
-    </html>
-  );
+  return <RootLayoutWrapper>{children}</RootLayoutWrapper>;
 }
