+++ apps/web/src/app/main/[serverId]/ServerProfile/page.tsx
@@ -1,14 +1,5 @@
-import {
-  ServerProfilePage,
-  projectSidebarList,
-  userList,
-} from '@meezy/ui/client';
+import { ServerProfileWrapper } from '@meezy/ui/client';
 
 export default function ServerProfile() {
-  return (
-    <ServerProfilePage
-      userList={userList}
-      projectSidebarList={projectSidebarList}
-    />
-  );
+  return <ServerProfileWrapper />;
 }
