+++ apps/web/src/app/main/[serverId]/page.tsx
@@ -1,5 +1,5 @@
-import { MainRoomClient } from '@meezy/ui/client';
+import { MainRoomWrapper } from '@meezy/ui/client';
 
 export default function Page() {
-  return <MainRoomClient />;
+  return <MainRoomWrapper />;
 }
