+++ libs/ui/src/lib/components/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 export * from './Header';
 export * from './CalendarMockup';
 export * from './DashboardCard';
