+++ libs/ui/src/lib/pages/index.ts
@@ -1,8 +1,12 @@
+'use client';
+
 export * from './MainRoomPage';
 export * from './FeedbackPage';
 export * from './SummaryPage';
 export * from './ChatRoomPage';
 export * from './ServerProfilePage';
 export * from './LoginPage';
-export * from './MainRoomClient';
+export * from './wrapper/pages/MainRoomWrapper';
 export * from './SignUpPage';
+export * from './wrapper/pages/ServerProfileWrapper';
+export * from './wrapper';
