+++ libs/ui/src/lib/pages/wrapper/index.ts
@@ -1,0 +1,2 @@
+export * from './layout';
+export * from './pages';
