+++ libs/ui/src/lib/pages/wrapper/layout/RootLayoutWrapper.tsx
@@ -1,0 +1,30 @@
+'use client';
+
+import './global.css';
+import {
+  ServerCreateProvider,
+  ServerJoinedTeamProvider,
+  ServerStateProvider,
+  ServerModalProvider,
+} from '../../../../context';
+
+export const metadata = {
+  title: 'Welcome to web',
+  description: 'Generated by create-nx-workspace',
+};
+
+export function RootLayoutWrapper({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang="en">
+      <body>
+        <ServerCreateProvider>
+          <ServerJoinedTeamProvider>
+            <ServerModalProvider>
+              <ServerStateProvider>{children}</ServerStateProvider>
+            </ServerModalProvider>
+          </ServerJoinedTeamProvider>
+        </ServerCreateProvider>
+      </body>
+    </html>
+  );
+}
