+++ libs/ui/src/lib/pages/wrapper/pages/MainRoomWrapper.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { MainRoomPage } from './MainRoomPage';
-import { projectSidebarList } from '../../context';
+import { MainRoomPage } from '../../MainRoomPage';
+import { projectSidebarList } from '../../../../context';
 import { useParams } from 'next/navigation';
 
-export function MainRoomClient() {
+export function MainRoomWrapper() {
   const params = useParams();
   const serverId = Number(params.serverId);
