+++ libs/ui/src/lib/pages/wrapper/pages/ServerProfileWrapper.tsx
@@ -1,0 +1,11 @@
+import { ServerProfilePage } from '../../ServerProfilePage';
+import { projectSidebarList, userList } from '../../../../context';
+
+export function ServerProfileWrapper() {
+  return (
+    <ServerProfilePage
+      userList={userList}
+      projectSidebarList={projectSidebarList}
+    />
+  );
+}
