+++ libs/ui/src/lib/pages/wrapper/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './MainPageWrapper';
+export * from './MainRoomWrapper';
+export * from './ServerProfileWrapper';
